Move event_bus service to ServiceProvider
